==> pipeline/00-preprocess/00_eng_a_extract_conversations.ipynb <==
import pandas as pd
df = pd.read_csv('../../data/raw/twitter/english/FolloweeIDs2_tweets_df_AugustPull.csv', dtype={'tweet_id': 'str', 'reply_to_id': 'str', 'original_user_id': 'str'})

conversations = []

df.head()

full_tweet_ids = set(df[df["reply_to_id"].isnull()]["tweet_id"])
first_order_replies_index = df[df["reply_to_id"].isin(full_tweet_ids)].index
full_tweet_ids_index_hashmap = {tweet["tweet_id"]: idx for idx, tweet in df[df["reply_to_id"].isnull()].iterrows()}

# collect first-order replies in the conversation format from above (= list[list[dict]])
first_order_replies = df.loc[first_order_replies_index]
for idx, row in first_order_replies.iterrows():
    conversation = []
    # find the original tweet
    original_tweet = df.iloc[full_tweet_ids_index_hashmap[row["reply_to_id"]]].to_dict()
    conversation.append(original_tweet)
    # add the reply
    conversation.append(row.to_dict())
    conversations.append(conversation)

with open('../../data/intermediate/english_conversations.pkl', 'wb') as f:
    pickle.dump(conversations, f)
